pyproject.toml
[tool.poetry]
name = "case-infinity"
version = "0.1.0"
description = "Project for Infinity Vision's technical case"
authors = ["Vinicius Peixoto <vinicius.opeixoto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
pydantic = "^2.8.2"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
scipy = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
blue = "^0.9.1"
ruff = "^0.5.7"
isort = "^5.13.2"

[tool.ruff]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
split_on_trailing_comma = "true"

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check . && blue --check . --diff && isort --check . --diff'
format = 'blue . && isort .'
pre_test = 'task lint'
test = 'pytest -s -x --cov=trees --cov-config=".coveragerc" -vv'
post_test = 'coverage html -d trees_html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
